1.4.33.58

+++ SolastaUnfinishedBusiness/Displays/EncountersDisplay.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text.RegularExpressions;
 using JetBrains.Annotations;
-using SolastaUnfinishedBusiness.Api.Infrastructure;
+using SolastaUnfinishedBusiness.Api.LanguageExtensions;
 using SolastaUnfinishedBusiness.Api.ModKit;
 using SolastaUnfinishedBusiness.Models;
 
@@ -33,10 +33,10 @@ internal static class EncountersDisplay
 
         using (UI.HorizontalScope())
         {
-            UI.ActionButton(actionText.Bold().Red(), action, UI.Width(30));
-            UI.Label($"{hero.Name} {hero.SurName}".Orange().Bold(), UI.Width(240));
+            UI.ActionButton(actionText.Bold().Red(), action, UI.Width((float)30));
+            UI.Label($"{hero.Name} {hero.SurName}".Orange().Bold(), UI.Width((float)240));
             UI.Label($"{hero.RaceDefinition.FormatTitle()} {hero.ClassesHistory[0].FormatTitle()}".White(),
-                UI.Width(120));
+                UI.Width((float)120));
 
             var attributesLabel = _showAttributes ? "" : "Attributes";
 
@@ -45,17 +45,17 @@ internal static class EncountersDisplay
             if (_showAttributes)
             {
                 UI.Label($"Str: {hero.GetAttribute(AttributeDefinitions.Strength).CurrentValue:0#}".White(),
-                    UI.Width(48));
+                    UI.Width((float)48));
                 UI.Label($"Dex: {hero.GetAttribute(AttributeDefinitions.Dexterity).CurrentValue:0#}".White(),
-                    UI.Width(48));
+                    UI.Width((float)48));
                 UI.Label($"Con: {hero.GetAttribute(AttributeDefinitions.Constitution).CurrentValue:0#}".Khaki(),
-                    UI.Width(48));
+                    UI.Width((float)48));
                 UI.Label($"Int: {hero.GetAttribute(AttributeDefinitions.Intelligence).CurrentValue:0#}".Khaki(),
-                    UI.Width(48));
+                    UI.Width((float)48));
                 UI.Label($"Wis: {hero.GetAttribute(AttributeDefinitions.Wisdom).CurrentValue:0#}".White(),
-                    UI.Width(48));
+                    UI.Width((float)48));
                 UI.Label($"Cha: {hero.GetAttribute(AttributeDefinitions.Charisma).CurrentValue:0#}".Khaki(),
-                    UI.Width(48));
+                    UI.Width((float)48));
             }
 
             var statsLabel = _showStats ? "" : "Stats";
@@ -65,12 +65,13 @@ internal static class EncountersDisplay
             if (_showStats)
             {
                 UI.Label($"AC: {hero.GetAttribute(AttributeDefinitions.ArmorClass).CurrentValue:0#}".White(),
-                    UI.Width(48));
-                UI.Label($"HD: {hero.MaxHitDiceCount():0#}{hero.MainHitDie}".Khaki(), UI.Width(72));
+                    UI.Width((float)48));
+                UI.Label($"HD: {hero.MaxHitDiceCount():0#}{hero.MainHitDie}".Khaki(), UI.Width((float)72));
                 UI.Label($"XP: {hero.GetAttribute(AttributeDefinitions.Experience).CurrentValue}".White(),
-                    UI.Width(72));
-                UI.Label($"LV: {hero.GetAttribute(AttributeDefinitions.CharacterLevel).CurrentValue:0#}".White(),
-                    UI.Width(48));
+                    UI.Width((float)72));
+                UI.Label(
+                    $"LV: {hero.GetAttribute(AttributeDefinitions.CharacterLevel).CurrentValue:0#}".White(),
+                    UI.Width((float)48));
             }
 
             CurrentItemsHeroes.TryGetValue(hero, out flip);
@@ -92,7 +93,7 @@ internal static class EncountersDisplay
         {
             using (UI.HorizontalScope())
             {
-                UI.Space(30);
+                UI.Space((float)30);
                 UI.Label("Inventory".Bold().Cyan());
             }
 
@@ -105,8 +106,8 @@ internal static class EncountersDisplay
 
                 using (UI.HorizontalScope())
                 {
-                    UI.Space(60);
-                    UI.Label(slot.EquipedItem.ItemDefinition.FormatTitle(), UI.Width(192));
+                    UI.Space((float)60);
+                    UI.Label(slot.EquipedItem.ItemDefinition.FormatTitle(), UI.Width((float)192));
                 }
             }
         }
@@ -121,9 +122,9 @@ internal static class EncountersDisplay
 
         using (UI.HorizontalScope())
         {
-            UI.ActionButton(actionText.Bold().Red(), action, UI.Width(30));
-            UI.Label($"{monsterDefinition.FormatTitle()}".Orange().Bold(), UI.Width(240));
-            UI.Label($"{SplitCamelCase(monsterDefinition.Alignment)}".White(), UI.Width(120));
+            UI.ActionButton(actionText.Bold().Red(), action, UI.Width((float)30));
+            UI.Label($"{monsterDefinition.FormatTitle()}".Orange().Bold(), UI.Width((float)240));
+            UI.Label($"{SplitCamelCase(monsterDefinition.Alignment)}".White(), UI.Width((float)120));
 
             var attributesLabel = _showAttributes ? "" : "Attributes";
 
@@ -131,12 +132,12 @@ internal static class EncountersDisplay
 
             if (_showAttributes)
             {
-                UI.Label($"Str: {monsterDefinition.AbilityScores[0]:0#}".White(), UI.Width(48));
-                UI.Label($"Dex: {monsterDefinition.AbilityScores[1]:0#}".Khaki(), UI.Width(48));
-                UI.Label($"Con: {monsterDefinition.AbilityScores[2]:0#}".White(), UI.Width(48));
-                UI.Label($"Int: {monsterDefinition.AbilityScores[3]:0#}".Khaki(), UI.Width(48));
-                UI.Label($"Wis: {monsterDefinition.AbilityScores[4]:0#}".White(), UI.Width(48));
-                UI.Label($"Cha: {monsterDefinition.AbilityScores[5]:0#}".Khaki(), UI.Width(48));
+                UI.Label($"Str: {monsterDefinition.AbilityScores[0]:0#}".White(), UI.Width((float)48));
+                UI.Label($"Dex: {monsterDefinition.AbilityScores[1]:0#}".Khaki(), UI.Width((float)48));
+                UI.Label($"Con: {monsterDefinition.AbilityScores[2]:0#}".White(), UI.Width((float)48));
+                UI.Label($"Int: {monsterDefinition.AbilityScores[3]:0#}".Khaki(), UI.Width((float)48));
+                UI.Label($"Wis: {monsterDefinition.AbilityScores[4]:0#}".White(), UI.Width((float)48));
+                UI.Label($"Cha: {monsterDefinition.AbilityScores[5]:0#}".Khaki(), UI.Width((float)48));
             }
 
             var statsLabel = _showStats ? "" : "Stats";
@@ -145,10 +146,10 @@ internal static class EncountersDisplay
 
             if (_showStats)
             {
-                UI.Label($"AC: {monsterDefinition.ArmorClass}".White(), UI.Width(48));
+                UI.Label($"AC: {monsterDefinition.ArmorClass}".White(), UI.Width((float)48));
                 UI.Label($"HD: {monsterDefinition.HitDice:0#}{monsterDefinition.HitDiceType}".Khaki(),
-                    UI.Width(72));
-                UI.Label($"CR: {monsterDefinition.ChallengeRating}".Khaki(), UI.Width(72));
+                    UI.Width((float)72));
+                UI.Label($"CR: {monsterDefinition.ChallengeRating}".Khaki(), UI.Width((float)72));
             }
 
             CurrentAttacksMonster.TryGetValue(monsterDefinition, out flip);
@@ -174,7 +175,7 @@ internal static class EncountersDisplay
             {
                 using (UI.HorizontalScope())
                 {
-                    UI.Space(30);
+                    UI.Space((float)30);
                     UI.Label("Features".Bold().Cyan());
                 }
 
@@ -189,8 +190,8 @@ internal static class EncountersDisplay
                             title = SplitCamelCase(feature.Name);
                         }
 
-                        UI.Space(60);
-                        UI.Label(title, UI.Width(192));
+                        UI.Space((float)60);
+                        UI.Label(title, UI.Width((float)192));
                     }
                 }
             }
@@ -208,7 +209,7 @@ internal static class EncountersDisplay
             {
                 using (UI.HorizontalScope())
                 {
-                    UI.Space(30);
+                    UI.Space((float)30);
                     UI.Label("Attacks".Bold().Cyan());
                 }
 
@@ -223,24 +224,24 @@ internal static class EncountersDisplay
                             title = SplitCamelCase(attackIteration.MonsterAttackDefinition.name);
                         }
 
-                        UI.Space(60);
+                        UI.Space((float)60);
                         UI.Label(title,
-                            UI.Width(192));
+                            UI.Width((float)192));
                         UI.Label($"action type: {attackIteration.MonsterAttackDefinition.ActionType}".Green(),
-                            UI.Width(120));
+                            UI.Width((float)120));
                         UI.Label($"reach: {attackIteration.MonsterAttackDefinition.ReachRange}".Green(),
-                            UI.Width(108));
+                            UI.Width((float)108));
                         UI.Label($"hit bonus: {attackIteration.MonsterAttackDefinition.ToHitBonus}".Green(),
-                            UI.Width(108));
+                            UI.Width((float)108));
                         UI.Label(
                             attackIteration.MonsterAttackDefinition.MaxUses < 0
                                 ? "max uses: inf".Green()
                                 : $"max uses: {attackIteration.MonsterAttackDefinition.MaxUses}".Green(),
-                            UI.Width(108));
+                            UI.Width((float)108));
 
                         if (attackIteration.MonsterAttackDefinition.Magical)
                         {
-                            UI.Label(TagsDefinitions.Magical.Green(), UI.Width(108));
+                            UI.Label(TagsDefinitions.Magical.Green(), UI.Width((float)108));
                         }
                     }
                 }
@@ -312,7 +313,7 @@ internal static class EncountersDisplay
                 UI.HStack(playerCharacters[index].Name, 1, () =>
                 {
                     if (UI.SelectionGrid(ref playerCharactersChoices[index], controllers, controllers.Length, 2,
-                            UI.Width(300)))
+                            UI.Width((float)300)))
                     {
                         PlayerControllerContext.PlayerCharactersChoices = playerCharactersChoices;
                     }
